Revert "stack::l2cap::tL2C_CONN_INFO Rename status => hci_status"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/1017af92-a80d-4b2a-ba38-0c127d278a3d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I22473b667f09e3cbd836f228465fe68c4ffc462c
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com