sdp: Increase SDP_MAX_CONNECTIONS parameter
SDP_MAX_CONNECTIONS doesn't mean the number of active connections, as
it also includes pending SDP requests. For example, when connecting
to a mouse, it's possible to have SDP requests for all of PNP, L2CAP,
PBAB, and HID. Even when they are resolved one by one, they still take
all currently available four control block slots.

It's been observed that on autotests we sometimes ran out of slots,
which is possible due to bad luck since sometimes PNP searches is done
multiple times. Theoretically, if this happens in practice, the user
can only forget and re-pair to force SDP to be carried out again.

This CL raises the number of SDP connection slots from 4 to 6.

Bug: 349143796
Test: m -j
Flag: EXEMPT, simple bugfix
Change-Id: Ibc4cdeb8be1a8278638303c887ed9cff7d70ee43