Revert "RAS: Ignore duplicate BTA_GATTC_SEARCH_CMPL_EVT"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/1ad94276-14a7-49ce-9c9b-7e6fc7790203). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I55711bc60c8ad245b033ffde5fefa2ed36093ad5
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com