Revert "::BleAdvertiserInterface: Reduce ambiguity using explici..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/7a0cd365-6751-4707-bfc3-86e049f66484). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I85c21ffb3c0d5be25f8e9a57b626b1f19f9844a6
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com