Revert "RAS: Ignore duplicate BTA_GATTC_SEARCH_CMPL_EVT"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/27fe2ee8-3eb4-4d31-b780-a20e07486337). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: If15c17565d496885bc01efd143f25c67b5adce6e
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com