Revert "add GMAP profile flag"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/68f4feac-f611-460e-8861-501604594a32). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 354220800
Change-Id: I9f55fe1ef11f86ba22af47c1cee50b6db72238b3
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com