Revert "Make Android Headtracker Service a restricted service"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/97ac75b0-3656-4cb1-a03e-a956af5aec15). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I978e897173bb1cefe260172dbcbadb4ea5981546
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com