Revert "InitFlags: Remove unused includes"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/7ac640ac-4971-45ed-a542-c28229fedf6e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I41607c9a54c45bc0b0df8e49cef02a2b8c87fd9c
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com