Revert "InitFlags: Remove unused includes"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/7e22b343-22dd-474a-8b8b-2b576551bd60). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Id18c04ca50361a6317b9f44acd0382e7e7084a51
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com