Revert "stack::l2cap Header clean up"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/7e22b343-22dd-474a-8b8b-2b576551bd60). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I6783845fba30c563f2736b1ea158d140e78ac9d3
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com