Revert "btif::dm::le_adv Properly type status as advertising status"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/ae170b00-5bee-4ff9-9725-2d5540c4eaa1). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I3a22e41f7c04131055b9411fc42eedb8813597d9
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com