Revert "floss: Modify A2dpCodecConfig::IsSinkCodecSupported to r..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/af3cd353-d06a-4f66-ac6c-dba6811c15ba). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I3f6c733987ce11764e2f040b506b727aa0356548
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com