Revert "hh: Queue input data when before receiving UHID_OPEN"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/1783275c-9c79-4bf0-a063-e6ad19641665). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I87026948a17b7ef881aa07c84f2c88c5318e4c97
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com