Revert "hh: run all uhid exchanges in uhid thread"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/1783275c-9c79-4bf0-a063-e6ad19641665). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ibb262436bb4d9e504572f592d5c6163a30232e5b
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com