Revert "hh: Queue input data when before receiving UHID_OPEN"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/1783275c-9c79-4bf0-a063-e6ad19641665). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I97c82790cbb056b386e6cdf738fba2af1e873874
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com