Revert "bta::ag::sdp_cback Use proper type tSDP_STATUS"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/eff18ada-5a29-41b3-889c-990ec8dab901). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ib127ea81c32118cae92875fb11143703459487b7
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com