SystemServer: mAdapterLock is overused
The lock is legacy. It comes from a times where there was multiples
thread that could update the mAdapter. This is no longer the case and
the only other thread that interact with mAdapter beside the main looper
is the binder thread.
The access to mAdapter from the binder thread is copying the value to
prevent asynchrone null set

Bug: 311772251
Test: atest CtsBluetoothTestCases
Test: atest pts-bot
Flag: Exempt refactor without logical change
Change-Id: I629c9992ef0d9849ad1301d45d127eaef7f1cb30