Revert "LeAudioService: getDevicesMatchingConnectionStates"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/cb884542-e529-4c34-a8e4-b96eee8c82d6). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I126c2b6f5f2ea9dae465d97e64d89e95615bae47
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com