Revert "LeAudioService: getDevicesMatchingConnectionStates"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/cb884542-e529-4c34-a8e4-b96eee8c82d6). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I0ff9f7462433a075259196fc792a917decb5aa3c
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com