Revert "Utils: Inline one line permission method"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/cb884542-e529-4c34-a8e4-b96eee8c82d6). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I2143b8535829ff8955608a0d1aa5c33ea6e2d27a
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com