Revert "Prevent un-necessary call to framework"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/cb884542-e529-4c34-a8e4-b96eee8c82d6). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I011d8ffc5e8052970faed84de2ee8e926d7ffbc0
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com