Revert "Bluetooth: fix RequiresPermission annotation"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/cb884542-e529-4c34-a8e4-b96eee8c82d6). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I185f506d660596d7611f83a622ce8b2faf832baf
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com