Revert "InitFlags: Remove unused includes"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/7a099c21-5eb0-4518-9409-5eb2a92bfe50). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I6c03e2618cf14d50ec4b995df22ba76befe1dc28
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com