Revert "floss: Add kBtSnoopLogModeKernel for btsnoop mode in kernel"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/5f571fce-7edf-4f98-8f01-61d8783ed641). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I44180e8f151d9648f5432c4d9b5b8069eacd1df9
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com