bta: le_audio: client: Tweak autoconnection to Coordinated Set Members
This changes default autoconnection logic to set members. Currently,
if just one device has been connected, the stack issued autoconnection
to remaining set members. The aim was to have all set members connected.

In case of Dynamic Set Members, we don't want to do that, because we
want to rely on Targeted Announcements. If the device becomes available
for the stream and starts Targeted Announcements, we will connect.
Otherwise, if the device is not available for stream, the connection
won't be created.

Bug: 357472821
Bug: 496639589
Flag: leaudio_unicast_dynamic_coordinated_sets
Test: atest bluetooth_le_audio_client_test
Change-Id: I0b01eedbc7decf477b398f5df9e3e45841792220