Add processing logic for Clcc in Connecting state to guarantee clccResponse.
Since hfp is not connected and not handle the msg "CLCC_RSP_TIMEOUT" in the connecting state machine, the clcc msg will be filtered out and not send the ClccResponse returned by telecom to stack, causing the SLC connection to fail and there is no response when you click the answer button on the car terminal. Therefore, we need add CLCC_RSP_TIMEOUT in connecting state guaranteed ClccResponse in any circumstances.

Bug:356514199
Tag: #feature
Test: manual
Change-Id: Id0b5899a0fc350ab3d9a5f5e26923c5851caccaa