Invoke MTU Callbacks on Failure
Previously, MTU configuration failures were not propagated
to the Java layer due to setting of `p_q_cmd` to NULL.

This change ensures that error callbacks are invoked
correctly when MTU configuration fails before a request is sent
by bypassing the NULL check.

It enables a test to confirm the behavior.

Bug: 356550596
Bug: 350622873

Test: atest GattClientTest
Change-Id: Ib5083a579ca881be4cd0d02ec4e20d265e15cf46