Revert "Adapter: use callback wrapper for quality report"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/b52b9576-c8b3-44b9-889d-0bda803d7504). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ic8d2a2a44bc05efc2ca9ab868b793745d4682282
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com