leaudio: Fix race when switching from CONVERSATIONAL to SOUNDEFFECT
Issue is triggered when speed reconfiguration flag is enabled but can
happen also with that.

The root cause is that SOUNDEFFECT is considered as a context which not
necessary changes the stream configuration. Stack decides about that by
looking into Streaming state, but it might happen that stream is still
in the STREAMING state and is waiting on remote to confirm RELEASING
state. In such a case, stack should move on with reconfiguration and not
keep previous stream configuration.

Without that patch, after stream is realeased, it will be reestablished
with Bidirectional CISes on the SOUNDEFFECT context type, which some
devices might not accept.

Bug: 356772046
Test: atest bluetooth_le_audio_client_test
Flag: EXEMPT, regression verified with unit tests
Change-Id: Ie993c8099902e605dc9ab0c5cafb438f743654f4