Revert "Configure always_fallback_to_available_device flag as bugfix"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/55e776aa-cd26-474b-bb33-30ea85416439). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: If011c9d4a244c05199e176136f2e3f0b91bd0403
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com