Revert "Configure always_fallback_to_available_device flag as bugfix"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/55e776aa-cd26-474b-bb33-30ea85416439). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I744ae8977a535e72370d6cae89e2cb98310431c5
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com