Revert "Configure fallback_when_wired_audio_disconnected flag as..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/55e776aa-cd26-474b-bb33-30ea85416439). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I67719c6be4ea41eba376b550af469afb3670cc9a
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com