Revert "Workaround breakage from LTO optimization"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/315d3cfb-ce40-4e43-bb04-fda3aeb2c77c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 356856433
Change-Id: Ib864c35480a165a4781f14d959a68cf8dbf497e2
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com