Revert "Dump BTA HH connection control blocks"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/34bdff6d-d7f3-498a-b364-3e27d34b5e36). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I1af37499ea70806e9551a0c4adac8e9700f8b1e0
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com