Revert "Early reservation of HID control block"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/34bdff6d-d7f3-498a-b364-3e27d34b5e36). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ied058fcc6ef0f5ffa78ee1fffe0f1e5f19eea40b
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com