Revert "Early reservation of HID control block"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/34bdff6d-d7f3-498a-b364-3e27d34b5e36). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I8a856ecb4f4937726c5da4f2d48ab447066f8b17
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com