Revert "Ignore empty pseudo and identity address"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/48e3212e-c8fb-423f-95b3-64fdb7773bb7). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Iad808c608dad20a04da54de4e12675a188242b42
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com