Revert "Stop calling app to know its state. Use Callback"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/015f820c-4a2a-4bc2-a281-8ff05354e9fa). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I41082b627af5789265a09ae0d267ae407966682d
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com