Revert "Fix "GATT Read Multiple Variable Response" builder"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/ae0a681a-67d3-481b-bfcc-4a03939b00df). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ie53fb75c90c6e251987ed9770bad44afa4bbac9b
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com