Revert "Make LDAC default quality mode configurable"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/457f34d2-394e-42a7-8334-e98629bebbf3). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 359410792
Change-Id: Ifc95d9af0aec1d824ddd2b4cd24fae0635db8b40
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com