RemoteDevices: Cancel delayed SDP UUID Intent if disconnected
The previously scheduled update of the UUID for the device is no longer
applicable as the device is currently disconnected. Maintaining this
message on the handler may result in inadvertent device connection
attempts, particularly if the device is in the process of disconnecting
or has already disconnected.

Bug: 324117559
Test: atest RemoteDevicesTest
Flag: Exempt covered with unit test.
Change-Id: Icb84097da814897de664ed7ff75ccdf0e48b7e79