Revert "Make LDAC default quality mode configurable"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e0084f3e-73f9-4b17-900e-01234d83d65e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 359410792
Change-Id: I9f01686799e3679decb8bcbbc092d66ebeaca4d3
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com