Revert "Remove unused flag"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/3f5b399e-9c73-448a-a93f-5641bd6a2a8e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ie60d16fffef9fb1c5653545040d005a5cfee6686
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com