Revert "[2/24] Canonical form tBTM_STATUS::BTM_BAD_VALUE_RET"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/01d7208e-8a98-430e-a252-d7ee3c2d937c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Iffea1eff16d8957ef7a91d08217f23cb4f013059
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com