Revert "stack::sdp [1/7] cpp lint fixes"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/01d7208e-8a98-430e-a252-d7ee3c2d937c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I9a75159c8142cab1e86f43d239113b453e1894a6
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com