Revert "[5/24] Canonical form tBTM_STATUS::BTM_CMD_STORED"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/460394ef-2977-4fe5-801e-c15de34a332f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Iec92299fa446e416cd3b81866e9f19f7ccdcaa67
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com