Revert "SystemServer: reformat flag dumpsys"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/3982b039-2624-4031-ac60-f44cdf588729). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 361486043
Change-Id: Ieea015bf5e2526ac0ec01567aa47535115e13c22
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com