Revert "SystemServer: reformat flag dumpsys"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/410f1a68-841d-4cdd-a4f5-d5c3ed105191). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 361478528
Change-Id: I35ce6bf11ac0346099b31d15b6620f374b59099f
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com