Revert "SystemServer: reformat flag dumpsys"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/d1fe6eaa-ccbd-4560-b55f-e1bd596a996e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 361478826
Change-Id: I94bb2abde048acef507b7376ebaddf2f7589ea12
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com