Revert "SystemServer: reformat flag dumpsys"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/412674c0-133e-48a0-85e0-597b5dcc0fb2). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 361479023
Change-Id: Id546bcf826aefad68a8253d72db9c44fe763328f
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com