Revert "SystemServer: reformat flag dumpsys"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/4de2990e-05da-4bb4-9b3b-1c71bd0045a9). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 361478742
Change-Id: I6253a56cb64be67651c5dc0bbcf9590e965ac679
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com