Revert "stack::btm Hide btm_api.h use get_btm_interface"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/ba3417f9-d0e6-4bfb-82ee-f52fc7b58002). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I02710eb3894e49baa20f59245192fe2e5df242a8
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com