Revert "SystemServer: reformat flag dumpsys"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/fcdcb133-01ac-4286-8c9c-3867abaa32a1). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 361508261
Change-Id: I9574ca3c9aea1ed1001409d539133f1a789a201a
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com