Revert "SystemServer: reformat flag dumpsys"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/60f1b220-c3ed-4ddc-baea-8d7e930adedb). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 361482888
Change-Id: I70e8dcbda1c8c0744cd9deb268296de833e9df74
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com