Floss: Add extra delay in cleanup before exit
In the final step of the current cleanup process, we are waiting for the
thread event to be disassociated, which is called after btif cleanup
finished. However, there are more cleanup functions needed to be done
after btif cleanup in |event_clean_up_stack|. These become undefined
behaviors since the process is going to be terminated by btadapterd.

This adds an extra delay to give the rest of the cleanup processes
some time to finish.

Bug: 304997914
Test: rmmod btusb to trigger sigterm and observe the extra delay
Test: mma -j
Tag: #floss
Change-Id: Id049a4b5c1192aae2b9a71ee1a7c394c9398f1c2