Revert "BtQualityReports cpplint"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/c0ef5896-8b82-453b-89cf-c4cb777dbb88). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I4eac780ec8702632a93549836ea9db09512ce125
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com