Correct SDP data element size descriptor for HIDD report descriptor
The length of the Report descriptor for HIDD is a unit16_t and limited by the API to 2048 bytes, is however written as an SDP data element with SIZE_IN_NEXT_BYTE, which leads to the element size being at most 255 bytes. The descriptor however is written fully to the SDP stream.

By using SIZE_IN_NEXT_WORD, we can accommodate for maximum size of 2048 bytes, which resolves these issues.

Bug: 362791956
Change-Id: Ie65e79d395c9f1db3adc0fef917a6c8db16bbe29
Test: manually