Revert "Wait for state callback before ending constructor"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/49823b7d-77f9-4e4d-91b1-41a2c3955116). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 362664845
Change-Id: I6ab998948b8c4fc28399437d7d4b9db295ce1b5c
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com