Revert "Wait for state callback before ending constructor"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/6b1f5d99-527f-48ea-8d3e-e9ec8f4220cc). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 362664845
Change-Id: Ic4606e93b47bbd62431157cad1bdadd2d207a304
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com