Revert "Revert "SystemServer: enable/disable from main thread ea..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/133a1601-f9b9-4026-b811-4507902301d8). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I1f990e5fa234626ca5f850afad93187c26357468
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com