DIS: don't attempt to connect if GATT isn't connected
We attempt to connect GATT first if it isn't connected, but there is
no handler for both the case the connection is successful and fail.
This blocks all future DIS requests since we mark this request as
ongoing, but it is never actually carried out.

For simplicity, just don't try to connect GATT if it isn't connected.
This prevents us from getting stuck, and we can implement the handlers
later.

Bug: 363093973
Test: m -j
Flag: EXEMPT, simple bugfix
Change-Id: I1780c74c02836acde134f263ff307a7ab1258706