Revert "Provide a flag library for rust code"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/1542998c-8ec5-4df8-ab02-2d10b0b0f292). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I4ac713ad4408d280ca3ee8703aa4244ccd92c3f8
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com