Revert "Provide a flag library for rust code"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/260168de-c591-482d-aabe-e8d51c839b33). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I2a64177f8d0651dfdb19323cac4eff88d7b19c79
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com