Revert "Update test node constructor to use random UUID"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/9d81f53a-82bd-4e0d-8e61-4ab062ca8c75). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I68b6007f65e31d3c4154200200781c05b6d83c5d
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com