Revert "Provide a flag library for rust code"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e4d0bf29-7e2a-4183-bd22-b48a1bebc4d8). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia2cbdcdb742a2f31538090403a95e370014a2fc6
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com