Revert "IsoManager: Unify protection against usage when not started"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/92703515-de5a-46b8-a50e-3c91e5982d88). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I0e152e3f1fb0553f0fccf7851b705dca5431cf5d
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com