Revert "Fix ClassCastException in BluetoothMediaBrowserService#d..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/7af50ff2-0ec6-4652-a6b6-70c32a6ac577). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia728db2c3a2ff81d7cd6f5255061157693011567
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com