Revert "Update test node constructor to use random UUID"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/7af50ff2-0ec6-4652-a6b6-70c32a6ac577). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ief6716dcb2584d7b39010fd84ac9ae5ffbff58f6
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com