Revert "Fix ClassCastException in BluetoothMediaBrowserService#d..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/b33b394a-db5d-4f4f-9ed0-6486e049d652). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I7d23f264e86cf30372c5802e73225466c527e41c
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com