Revert "Add DisableBtQualityReport"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e5841427-0e4e-4deb-a3a3-25330eed3f79). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ic11b22464a5d3570ee01592c1d44519d9ef5bcd5
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com