Revert "floss: Refactor btstack initialization order (3/4 Blueto..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/7bcd76cc-245a-4626-942d-6c5765fb8a49). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia255385b71be6181e43969e09bb9f978f22fa7e6
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com