Revert "Alarm: Create timerfd with TFD_NONBLOCK flag"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/544d43ea-fd60-47e3-bc32-3f1cfa58be27). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I73103b69dde55359d56750935e0d68005fa7a392
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com