Revert "[MTK][CONVERGENCE][ALPS07830065] Use 1M to response"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/bb9a26aa-9d1c-41aa-9a87-38a4aa52bfa3). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I3dc4787e089baa83327c6684f3ea1df858b82fde
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com