Revert "[MTK][CONVERGENCE][ALPS07830065] Use 1M to response"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e7b96737-7523-459c-81d7-e608155be5c3). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I1a898a7f9705638aaa9dbc88d1fc5744a7bee381
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com