Revert "Flag: Add close_hid_if_uhid_ready_too_slow"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/fd133c7c-957b-4c27-b494-5b24a4cd3621). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ibab5f263399f42b12ae5f71211ac3bd431a7dcc5
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com