SystemServer: unbind prior to setting state
calling mState.set() will let the caller of waitForState to return
already.
But when transitioning to stateOff, the caller expect the Bluetooth to
have completely finish it's cleanup and no longer be around.

By calling unbindAndFinish prior to setting the state, we reduce the
probability of this scenario

Test: atest ServiceBluetoothRoboTests
Test: atest ServiceBluetoothTests
Test: any code that toggle Bluetooth off and on
Flag: Exempt this target flakiness and is check threw intensive abtd
Bug: 361850050
Change-Id: If3a5e6e358e295a2f77619d5f1c82c69c950a6e2