Revert "Clean up flag update_inquiry_result_on_flag_change."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/ffc026fd-fac4-4ace-9def-dbee0e632777). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ic17c212a5c5a123ac1be05be2cf7cf3c65aec4c4
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com