Revert "Flag: Add close_hid_if_uhid_ready_too_slow"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/88f4fe06-8777-49e9-8c00-cd2eeed414bb). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I51b6026de462698d5fe7635ebda26363281268c3
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com