Revert "Alarm: Create a new constructor for nonwake alarm support"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/eb197c2d-a1f4-4145-9e2b-8da4264f2d3e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: If361692113468fa02fa7a9bc826d5eddd56ce5ba
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com