Revert "stack::btm::sco_hci - interface with non-Floss build"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/d4d08625-d5b1-4206-9173-5eb3376941d6). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Iaabe80c25a6bdeec9a0178117e341f41288b6ec2
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com