Revert "Alarm: Create a new constructor for nonwake alarm support"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/c872c529-4c04-4cf7-8b82-ec134f570304). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ifbe15d7514573064e640d88f06bb68e9a066cf79
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com