Revert "SystemServer: Manually intiate state transition"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/6acc6999-120a-482c-8fb3-a5174282a74c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I04b0f5dfe00e0295e5599c79c1791a61a2b8f8d3
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com