Revert "stack::btm::sco_hci - interface with non-Floss build"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/b74ff86e-6c35-4350-a66c-215429029bb5). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ice07ddadeaa532b50805861202286be2a2ff5f8e
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com