Prioritize strict address match over RPA resolution
btm_find_dev() finds the device record for the given address. It goes
through the list of all device records and returns record where
1. identity address matches the given address,
2. pseudo address matches the given address, or
3. IRK can resolve the given address.
This means that if an IRK from one of device in the list prior to target
device record is able to resolve the address, that device record will be
returned.
The updated logic now first compares the identity and pseudo address
from the device records. If the exact device record is not found, then
it tries to find the device record which can resolve the address.

Test: mmm packages/modules/Bluetooth
Flag: EXEMPT bugfix
Bug: 372533704
Change-Id: I8a87d7a47e024014b0a909e0c468b197c774be32