Revert "Only disable BtQualityReports once"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/f58d410f-b800-414d-bb89-c93876a25804). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I61553ba2ad4a3d2b3769b79a69e6c649d09eaac7
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com