Revert "SystemServer: Remove un-necessary mAdapterLock"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/b961b2d5-9064-4d8f-a3e6-26a3102677d4). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 366103445
Change-Id: I71affdaece7cd9c12144dca369e215d75c04fbf9
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com