Revert "stack::btm::sco_hci - interface with non-Floss build"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/54803d96-461d-49a9-a8ab-33b12cab483e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I2048e869ef4967b2ed8c4ad36d7f6ad04c17176c
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com