Revert "Alarm: Create a new constructor for nonwake alarm support"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/c017f2b3-2217-4c5d-ae29-121ffb285a5b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ieeaa3a60fb15b3b099b645514e68796ab4fcabb0
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com