Revert "[21/25] stack::rfcomm Use stack::l2cap::get_interface()"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/95c6f3a8-5c8b-4272-9a5c-b1ba93479a9e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ic7c977e6a6988d21a34a63dc2fb41736413443d2
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com