Remove restricted devices only in non-restricted mode
GD storage module does not know if the mode is restricted or not. It
always remves the restricted devices on startup.
This change moves the decision to remove restricted devices from the
storage to btif_storage module.
This change also links the device record created in restricted mode to
the user ID. This ensures that different guests cannot access each
other's devices.

Test: mmm packages/modules/Bluetooth
Test: Manual | Pair with a device in guest mode and restart the BT,
paired device must still be available in the guest mode
Flag: com.android.bluetooth.flags.guest_mode_bond
Bug: 346337854
Bug: 349882273

Change-Id: I909053c7ffc5f5ac2ff2b8fc3840f3b77c7b543a