HfpClientInterface: allow uni-directional audio stream on SCO
In some cases (e.g., on start-suspend transient states), there could be
only one direction of active audio stream, and either of the r/w is not
allowed.

In this CL, we capture the precise state of the active directions, and
make the streams appear to be active to the lower layers that assume
both are always active.

Test: verify BT audio works
Bug: 349290628
Flag: com::android::bluetooth::flags::is_sco_managed_by_audio and HFP

Change-Id: Ide1006b016c94960d5e869b5b48c85d3e3a7754f