Revert "Alarm: Create a new constructor for nonwake alarm support"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/878d0a26-7f5e-43c0-8cfe-8ce16f0b766f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia94bee55ceba663efedfd644b3cca0057dd1b23e
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com