Revert "Add flag to clear pairing state on early bonding failure"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/62af720b-965b-4363-8a6a-eeb0ed08095e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I0ee21a262cdb9eb0c48a3a551bc88a4de90ac2e9
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com