Revert "Alarm: Create a new constructor for nonwake alarm support"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/d8393f7d-a36b-4515-b367-b980a5b950fd). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I74e67d039673af320c1960696e674122ee836939
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com