Revert "pts-bot: skip failing HAP tests"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/c4814fb6-991c-4374-9bc9-39b12d1da4d4). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I125110437392313eb46292880b8f9d28346fcfac
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com