leaudio: Read Available Context Type after reconnect
As per PACs specification,
"3.5.1. Available Audio Contexts behavior
<snip>

If the characteristic value exposed for a client changes when not in a connection,
and the value of the Client Characteristic Configuration descriptor is configured
for notifications, the server shall notify the new characteristic value when
reconnecting to a bonded client.
"

However it was observed that some audio servers are not doing that, which
leads to unexpected behavior.

This value changes quite often when multipoint scenario is used.

For this reason, and since Android already reads ASCS Control Point CCC
descriptor on reconnection, Android will read this value after
reconnection.
Since for EATT supported devices, Read Multiple Variable Len is going to
be used, there is almost no cost for this read.

This patch also moves handling of ASCS CTP CCC descriptor to other
place - this is mechanical refactor.

Bug: 366062501
Test: atest bluetooth_le_audio_test
Flag: EXEMPT, regression test with unit tests, new test added
Change-Id: I71d058343ef4408ad3f730261ee26563728a8d4f