Floss: Remove delay to register callback in non-interactive mode
Only introduce delay on interactive mode. This is because we expect the
user to ensure the adapter interface is ready when they issue the
command in non-interactive mode. Otherwise, there will always have 1
second delay and in most of the case it is not needed.

Bug: 368387445
Tag: #floss
Test: mmm packages/modules/Bluetooth
Test: btclient --command "media info" and observe no delay
Test: btclient adapter enable in interactive mode and observe a delay
Flag: EXEMPT, Floss-only changes

Change-Id: Ica0d9146c1947c3cd9236b5d2541d8fc8c9bf3d8