floss: Connect to newly discovered profiles if it is user initiated
It's possible for some devices to announce a different subset of
profiles on the inquiry result, advertisement, and SDP. Currently our
"connect to all profiles" might not perform as expected because by the
time we finish connecting to all profiles, a new profile might be
discovered via another means.

Therefore, update the connection logic to always connect to newly
discovered profiles if "connect to all profiles" was invoked.

Bug: 352632969
Test: m -j
Test: autotest
Flag: EXEMPT, floss only changes
Change-Id: I68a12efab78af901eb1f5e35abeb9ac5da7db8ff