ActiveDeviceManager: stop audio when switched between ASHA and LE Audio
Depending if fallback device is already available or not, audio playback
was continued or not. This is unconsistent behavior for scenario when
same device is switching from ASHA to LE Audio profile. Now,
check if fallback device exists, and if it does, if the profile is
different. If the type changes, stop audio.

Bug: 368221990
Test: atest ActiveDeviceManagerTest
Flag: Exempt, feature not released yet
Change-Id: I290a795db89aa75fcffa5518f16acbe5712ecf1c