bluetooth: acquire wakelock when connection is complete
Acquire a wakelock when receving the connection success event. Wait for
a timeout to release the wakelock, so the bluetooth device has time to
send input events and wake up the system. If bluetooth resumes first,
also release the wakelock. The timeout is defined with sysprop. If the
timeout is zero, the request for wakelock is ignored.

Bug: 369686962
Bug: 366432079
Test: m -j, suspend system and wake up with mouse
Flag: adapter_suspend_mgmt
Change-Id: Ia6aab33857d4d7a2ddc7c5ff8e182ab42cb77825