Revert "Pandora: remove trendy ownership"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/944ebfef-fbf6-449a-bbd3-1a0c86cc9308). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I73099a5f288345388cafac92ae9c42d2292e9ad5
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com