Revert "Workaround breakage from LTO optimization"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/8caee421-88b3-4945-b55a-cac0d8e33704). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 370254314
Change-Id: I19a0579066eb31cbcab28eee7458e7f422266cad
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com