Add VFS browsing Cover Art support & set AVRCP 1.6 default
Cover art was implemented for browsing, but only when remote requested
Now Playing browsing. This was due to memory usage required to store images for
VFS browsing as the media players send an URI instead. We need to implement it,
and gate it behind a config option to enable as wanted to pass certification
tests. The implementation is already done in the native side.

We should also move to AVRCP 1.6 as default version and enable cover art.

Change-Id: Ib970e3add951bc7e9581e1e4b38b67d215828bb4