btif_sock_rfc: Cancel ongoing SDP on socket closed
When a socket is closed while the SDP is still ongoing, btif socket slot
would be cleaned up but BTA Jv would not, which makes all later RFCOMM
connection fail. The issue usually persists until the SDP completes, but
we also observed that the SDP complete event could never happen. This
patch shall mitigate the issue.

Bug: 331018524
Bug: 370916753
Test: atest BumbleBluetoothTests:android.bluetooth.RfcommTest#clientConnectToOpenServerSocketBondedInsecurePrematureClosure; Passed with flag enabled and failed with flag disabled
Change-Id: If2bdc0f0d3b5f9cfc48073f1788241fc296582de