Revert "SystemServer: test race condition on enable"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/b5c48fa3-6f72-4cbd-8489-7f66a7fde267). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 371054808
Change-Id: I110c1d02e0c94041bd1b580c08451194a68beaba
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com