Revert "SystemServer: test race condition on enable"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/acc00142-9d64-42b8-baac-d8e18fd19e9e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 371054730
Change-Id: Id2e023033cc05ffd7686126025738397d1392f3c
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com