Revert "SystemServer: Remove un-necessary synchronize"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/acc00142-9d64-42b8-baac-d8e18fd19e9e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 371054730
Change-Id: Iff4583ebd8d6f292d0249de4fdd6243ce6a4ac1e
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com