Revert "SystemServer: test race condition on enable"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/dae5d496-b1ab-4264-bef8-5b0ec3aa103a). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 371054855
Change-Id: I2a213ba8858ee60253e28333dec6bb874fdd7527
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com