Revert "Disable SWB for desktop if bluetooth.hfp.swb.supported=false"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/2cdfd67c-25fa-4ec1-a062-c71fde8740b5). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 371171824
Change-Id: Icdbff091c9ed9c7d319a6795b7d22e48b8e45192
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com