Revert "system/btif: Fix -Wmissing-prototypes errors"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/c2675d92-ef1c-4d80-83e1-d856b0ea46ff). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 371445480
Bug: 371453061
Change-Id: I2dbd15149450a16a4a639b76fcf2969d43560a85
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com