floss: dbus_projection: Allow other attributes on top of dbus_method
The current logic in dbus_projection is very strict when checking for
the dbus_method attribute, specifically it doesn't allow other
attributes at all. This might come as surprising because even an
unsuspecting function comment actually counts as an attribute. When
parsing failure occurs, the macro treats as if it's not a dbus function
so no error is reported.

Change the logic to allow multiple attributes. For now those
attributes are ignored - if more support is required one needs to add
their own implementation still.

Bug: 371534263
Test: m -j
Test: btclient advertise on, verify doesn't crash

Change-Id: I26d2595b0c536a2dbd0ae505f6234f881e2199f7