Fix background connection
The `cif_set` should have been used instead of `cif_mask`
when dynamic allocation is enabled.

Using `cif_mask` prevented background connections
from being established when the flag was turned on.

This change corrects the function call and adds a test
for background GATT connections to ensure
the issue is resolved and prevent future regressions.

Bug: 371617855
Bug: 348559823
Test: atest GattClientTest

Change-Id: I5e8fba7e3b133475fe2e98581209145b75d3fb41