interop: extended truncated report map for specific device
The HID specification allows report maps to exceed 512 bytes. However, the Bluetooth specification requires all GATT attributes to be 512 bytes or less. This creates an interoperability issue when a device manufacturer implements a HID report map as a single GATT characteristic. This change handles this edge case by extending the report map after receiving the initial 512 bytes. This ensures compatibility with both HID and Bluetooth standards.

As this workaround is currently implemented for one specific device, the extra bytes of the report map exist as static const arrays; if more devices need such workarounds, it is suggested to store the report map in a separate file which can be loaded from the file system.

Bug: 332761737
Tag: #floss
Test: mmm packages/modules/Bluetooth
Test: Manual test on Brya, Brydge keyboard working
Flag: EXEMPT, interop change and apply only to one and non-functioning device
Change-Id: If9eb2e5e2828d0dea17f486207421525ffcccffc