Revert "Remove usage of memset"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/fe63320f-062c-4a60-bc2b-684fb0184582). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I44e425351f3ee904522f4747ec38cfe0b82ca5d1
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com