Revert "Bumble Java Android Headtracker Test Infra Changes"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/1ebf38f6-1951-410f-aa77-9d176eda2d15). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 372518178
Change-Id: I86db5f262e7004e47d16e563bcdbbb5cdbbac7b9
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com