Revert "Modified HidHostDualModeTest as per new hid rpc call"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/1ebf38f6-1951-410f-aa77-9d176eda2d15). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 372518178
Change-Id: Icc06027ec94501ff6fd6b84fceb7dd2d9ba7cd7c
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com