Revert "Modified HidHostTest as per new hid rpc call"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/1ebf38f6-1951-410f-aa77-9d176eda2d15). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 372518178
Change-Id: I0efa85999e4ff877f4dabe2b1880fa60e7bce67a
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com