Revert "HeadsetClient Test: Truth & intent matcher"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/74a6193d-35d7-4bf2-a434-26516ff2b9b4). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I64185867ef34c1d85c940b9fa9b1c9a349d0133c
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com