Revert "Remove unused parameters from bt_interface_t::init"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/7226d426-75ee-473c-acd4-cf33151a51c0). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I0fda9bf637f9532358c4922ac0b0c996613a9ec7
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com