Revert "leaudio: Do not update local metadata if not needed"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/fa39efbe-a76a-43c8-8984-61443b7758da). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Iec61046f3ee0eb4abb718932f624abda58401976
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com