GATT: Hold a wakelock while gatt_rsp_timer_ent is scheduled
On TV, the device might suspend while waiting for a GATT
response. If the suspend lasts for more than 30 seconds,
this would automatically disconnect the device upon
resume, forcing the user to manually re-connect the
device. Holding a wakelock for that length of time
should be reasonable since we expect most GATT
requests to be responded to pretty quickly.

Bug: 343826774
Test: m -j128 com.google.android.btservices & teamfood testing
Change-Id: I3e81e1eb6a22fe789a9ebdafb9f5013d8d4000b2