Revert "leaudio: Do not update local metadata if not needed"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/30c2671d-7127-406f-a60f-81bb3b50044a). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Iba617f1048d25c97642f5b713cb8ca46f06be516
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com