Revert "Update PDL for the CS HCI according to the released BT c..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/668b5d62-ea20-4b39-93b7-8c36f402f437). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ib5f61c3ebc177ceb4557c065f33bc61230805c10
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com