Revert "system/btif: Enforce -Wmissing-prototypes"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/d7d133e1-fb9c-4f46-b82d-9cbb3af6d364). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Idb26333677ccb770571db41e5e1b49a2096f5145
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com