Revert "system/bta/pan: Enforce -Wmissing-prototype"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/d7d133e1-fb9c-4f46-b82d-9cbb3af6d364). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Id90340f23a3627cb3e6a457abc1e978fb9126166
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com