Revert "system/bta/{ag,ar}: Enforce -Wmissing-prototypes"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/d7d133e1-fb9c-4f46-b82d-9cbb3af6d364). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I2928e28cd28bcb8101246de0079cce141c6923d8
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com