Revert "Fix typo in A2dpService.java"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/4072822e-1471-4ae0-b5ad-a3f38920e9f0). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 373154957
Change-Id: I6718b45caaf4044861b733e68d5bc6a5d398bc07
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com